requirements.txt
alembic==1.4.2
bcrypt==3.1.7
blinker==1.4
cachelib==0.1.1
cffi==1.14.1
click==7.1.2
cryptography==3.0
dnspython==2.0.0
email-validator==1.1.1
flake8==3.8.3
Flask==1.1.2
Flask-Cors==3.0.8
Flask-Login==0.5.0
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-Session==0.3.2
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==2.10
importlib-metadata==1.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
mccabe==0.6.1
mysql==0.0.2
mysqlclient==2.0.1
passlib==1.7.2
protobuf==3.12.2
psycopg2==2.8.5
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
six==1.15.0
SQLAlchemy==1.3.18
SQLAlchemy-serializer==1.3.4.2
Werkzeug==1.0.1
WTForms==2.3.1
zipp==3.1.0
